Extract window helpers
